requirements.txt
alembic==1.14.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
Flask==2.2.5
flask-login==0.6.2
flask-migrate==4.1.0
flask-sqlalchemy==3.0.1
Flask-WTF==1.2.2
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-cloud-vision==3.4.5
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.63.0rc1
gunicorn==21.2.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pillow==10.2.0
pluggy==1.5.0
proto-plus==1.26.0
protobuf==4.25.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
SQLAlchemy==1.4.48
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
werkzeug==2.2.3
WTForms==3.2.1
